+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/SearchHighlighter.java
@@ -30,6 +30,7 @@ import is.codion.swing.common.ui.key.KeyEvents;
 
 import javax.swing.JPopupMenu;
 import javax.swing.JTextField;
+import javax.swing.JViewport;
 import javax.swing.text.BadLocationException;
 import javax.swing.text.DefaultHighlighter;
 import javax.swing.text.DefaultHighlighter.DefaultHighlightPainter;
@@ -37,6 +38,7 @@ import javax.swing.text.Highlighter;
 import javax.swing.text.Highlighter.HighlightPainter;
 import javax.swing.text.JTextComponent;
 import java.awt.Color;
+import java.awt.Point;
 import java.awt.Rectangle;
 import java.util.ArrayList;
 import java.util.List;
@@ -45,6 +47,7 @@ import java.util.regex.Pattern;
 
 import static is.codion.common.resource.MessageBundle.messageBundle;
 import static is.codion.swing.common.ui.Colors.darker;
+import static is.codion.swing.common.ui.Utilities.parentOfType;
 import static is.codion.swing.common.ui.control.Control.commandControl;
 import static java.awt.event.KeyEvent.*;
 import static java.util.Objects.requireNonNull;
@@ -52,7 +55,7 @@ import static java.util.ResourceBundle.getBundle;
 
 /**
  * Highlights search results in a JTextComponent.<br>
- * Instantiate via the {@link SearchHighlighter#searchHighlighter(JTextComponent)} factory method.
+ * Instantiate via {@link SearchHighlighter#builder(JTextComponent)}.
  */
 public final class SearchHighlighter {
 
@@ -63,21 +66,24 @@ public final class SearchHighlighter {
 	private final Value<String> searchStringValue = Value.nonNull("")
 					.listener(this::searchAndHighlightResults)
 					.build();
-	private final State caseSensitiveState = State.builder()
-					.listener(this::searchAndHighlightResults)
-					.build();
+	private final State caseSensitiveState;
 	private final Highlighter highlighter = new DefaultHighlighter();
 	private final List<MatchPosition> searchTextPositions = new ArrayList<>();
 	private final Value<Integer> currentSearchTextPositionIndex = Value.value();
 	private final Value<Integer> selectedSearchTextPosition = Value.value();
+	private final ScrollToRatio scrollToRatio;
 
 	private HighlightPainter highlightPainter;
-	private HighlightPainter highlightSelectedPainter;
+	private HighlightPainter selectedHightlightPainter;
 
-	private SearchHighlighter(JTextComponent textComponent) {
-		this.textComponent = requireNonNull(textComponent);
-		highlightPainter = new DefaultHighlightPainter(darker(textComponent.getSelectionColor()));
-		highlightSelectedPainter = new DefaultHighlightPainter(textComponent.getSelectionColor());
+	private SearchHighlighter(DefaultBuilder builder) {
+		this.textComponent = builder.textComponent;
+		this.scrollToRatio = new ScrollToRatio(builder.scrollYRatio, builder.scrollXRatio);
+		this.caseSensitiveState = State.builder(builder.caseSensitive)
+						.listener(this::searchAndHighlightResults)
+						.build();
+		highlightPainter = new DefaultHighlightPainter(builder.selectedHighlightColor);
+		selectedHightlightPainter = new DefaultHighlightPainter(builder.highlightColor);
 		textComponent.setHighlighter(highlighter);
 		bindEvents(textComponent);
 	}
@@ -107,8 +113,8 @@ public final class SearchHighlighter {
 	/**
 	 * @param color the color to use when highlighting the selected search result.
 	 */
-	public void highlightSelectedColor(Color color) {
-		highlightSelectedPainter = new DefaultHighlightPainter(requireNonNull(color));
+	public void selectedHighlightColor(Color color) {
+		selectedHightlightPainter = new DefaultHighlightPainter(requireNonNull(color));
 		searchAndHighlightResults();
 	}
 
@@ -133,12 +139,11 @@ public final class SearchHighlighter {
 	}
 
 	/**
-	 * Instantiates a new search highlighter for the given text component.
-	 * @param textComponent the text component to search
-	 * @return a new {@link SearchHighlighter} for the given component
+	 * @param textComponent the text component
+	 * @return a new {@link Builder} instance
 	 */
-	public static SearchHighlighter searchHighlighter(JTextComponent textComponent) {
-		return new SearchHighlighter(textComponent);
+	public static Builder builder(JTextComponent textComponent) {
+		return new DefaultBuilder(requireNonNull(textComponent));
 	}
 
 	/**
@@ -208,7 +213,7 @@ public final class SearchHighlighter {
 		selectedSearchTextPosition.set(matchPosition.start);
 		try {
 			highlighter.removeHighlight(matchPosition.highlightTag);
-			matchPosition.highlightTag = highlighter.addHighlight(matchPosition.start, matchPosition.end, highlightSelectedPainter);
+			matchPosition.highlightTag = highlighter.addHighlight(matchPosition.start, matchPosition.end, selectedHightlightPainter);
 		}
 		catch (BadLocationException e) {
 			throw new RuntimeException(e);
@@ -232,17 +237,44 @@ public final class SearchHighlighter {
 		textComponent.getDocument().addDocumentListener((DocumentAdapter) e -> searchAndHighlightResults());
 		selectedSearchTextPosition.addConsumer(selectedSearchPosition -> {
 			if (selectedSearchPosition != null) {
-				try {
-					Rectangle rect = textComponent.modelToView(selectedSearchPosition);
-					if (rect != null) {
-						textComponent.scrollRectToVisible(rect);
-					}
-				}
-				catch (BadLocationException e) {
-					throw new RuntimeException(e);
-				}
+				scrollToPosition(textComponent, selectedSearchPosition);
 			}
 		});
+	}
+
+	private void scrollToPosition(JTextComponent component, int position) {
+		JViewport viewport = parentOfType(JViewport.class, component);
+		try {
+			Rectangle view = component.modelToView(position);
+			if (viewport != null && view != null) {
+				viewport.setViewPosition(viewPosition(viewport, view));
+			}
+		}
+		catch (BadLocationException e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	private Point viewPosition(JViewport viewport, Rectangle view) {
+		return new Point(locationX(viewport, view), locationY(viewport, view));
+	}
+
+	private int locationY(JViewport viewport, Rectangle view) {
+		int extentHeight = viewport.getExtentSize().height;
+		int viewHeight = viewport.getViewSize().height;
+		int yPosition = (int) Math.max(0, view.y - ((extentHeight - view.height) * scrollToRatio.scrollYRatio));
+		yPosition = Math.min(yPosition, viewHeight - extentHeight);
+
+		return yPosition;
+	}
+
+	private int locationX(JViewport viewport, Rectangle view) {
+		int extentWidth = viewport.getExtentSize().width;
+		int viewWidth = viewport.getViewSize().width;
+		int yPosition = (int) Math.max(0, view.x - ((extentWidth - view.width) * scrollToRatio.scrollXRatio));
+		yPosition = Math.min(yPosition, viewWidth - extentWidth);
+
+		return yPosition;
 	}
 
 	private static JPopupMenu createPopupMenu(ToggleControl caseSensitiveControl) {
@@ -252,6 +284,119 @@ public final class SearchHighlighter {
 						.build());
 
 		return popupMenu;
+	}
+
+	/**
+	 * Builds a {@link SearchHighlighter}.
+	 */
+	public interface Builder {
+
+		/**
+		 * @param highlightColor the highlight color
+		 * @return this builder
+		 */
+		Builder highlightColor(Color highlightColor);
+
+		/**
+		 * @param selectedHighlightColor the selected highlight color
+		 * @return this builder
+		 */
+		Builder selectedHighlightColor(Color selectedHighlightColor);
+
+		/**
+		 * @param caseSensitive true if the search should be case sensitive, default tue
+		 * @return this builder
+		 */
+		Builder caseSensitive(boolean caseSensitive);
+
+		/**
+		 * @param scrollYRatio specifies the Y axis scroll ratio when scrolling to a search result, 0 being at the top and 1 at the bottom, default 0.5
+		 * @return this builder
+		 * @throws IllegalArgumentException in case the value is not between 0 and 1
+		 */
+		Builder scrollYRatio(double scrollYRatio);
+
+		/**
+		 * @param scrollXRatio specifies the X axis scroll ratio when scrolling to a search result,
+		 * 0 being all the way to the left and 1 all the way to the right, default 0.5
+		 * @return this builder
+		 * @throws IllegalArgumentException in case the value is not between 0 and 1
+		 */
+		Builder scrollXRatio(double scrollXRatio);
+
+		/**
+		 * @return a new {@link SearchHighlighter}
+		 */
+		SearchHighlighter build();
+	}
+
+	private static final class DefaultBuilder implements Builder {
+
+		private final JTextComponent textComponent;
+
+		private Color highlightColor;
+		private Color selectedHighlightColor;
+		private boolean caseSensitive = true;
+		private double scrollYRatio = 0.5;
+		private double scrollXRatio = 0.5;
+
+		private DefaultBuilder(JTextComponent textComponent) {
+			this.textComponent = textComponent;
+			this.highlightColor = textComponent.getSelectionColor();
+			this.selectedHighlightColor = darker(textComponent.getSelectionColor());
+		}
+
+		@Override
+		public Builder highlightColor(Color highlightColor) {
+			this.highlightColor = requireNonNull(highlightColor);
+			return this;
+		}
+
+		@Override
+		public Builder selectedHighlightColor(Color selectedHighlightColor) {
+			this.selectedHighlightColor = requireNonNull(selectedHighlightColor);
+			return this;
+		}
+
+		@Override
+		public Builder caseSensitive(boolean caseSensitive) {
+			this.caseSensitive = caseSensitive;
+			return this;
+		}
+
+		@Override
+		public Builder scrollYRatio(double scrollYRatio) {
+			if (scrollYRatio < 0 || scrollYRatio > 1) {
+				throw new IllegalArgumentException("scrollYRatio must be between 0 and 1");
+			}
+			this.scrollYRatio = scrollYRatio;
+			return this;
+		}
+
+		@Override
+		public Builder scrollXRatio(double scrollXRatio) {
+			if (scrollXRatio < 0 || scrollXRatio > 1) {
+				throw new IllegalArgumentException("scrollXRatio must be between 0 and 1");
+			}
+			this.scrollXRatio = scrollXRatio;
+			return this;
+		}
+
+		@Override
+		public SearchHighlighter build() {
+			return new SearchHighlighter(this);
+		}
+	}
+
+	private static final class ScrollToRatio {
+
+		private final double scrollYRatio;
+		private final double scrollXRatio;
+
+		private ScrollToRatio(double scrollYRatio, double scrollXRatio) {
+			this.scrollYRatio = scrollYRatio;
+			this.scrollXRatio = scrollXRatio;
+		}
 	}
 
 	private static final class MatchPosition {